Fixed bug of flickering cursor:
Problem occurred because we use MouseEnter, MouseWithin, MouseLeave to update cursor and change it. 

However, in the new version, AdjustCursor was changed to AdjustMouse and AdjustMouse was defaulted to update cursor to arrow.

Thus, code warrior continually called AdjustMouseSelf and reupdated our arrow. Thus we had a fight between the two and it flickered back and forth between the arrow and the hand, or the arrow took total control. 

We solved this problem by overriding AdjustMouseSelf with the call to update cursor as we had previously. 

(MouseEnter etc called CheckCursor in the cursor manager and now we have AdjustMouseSelf call CheckCursor in cursor manager).

These changes were done by Elizabeth Hamon and Eric Kidd. Commits made with Eric's approval.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@27075 b24aca02-f418-0410-9c49-835e4aa915c3