Made changes in order to show and hide the menu bar correctly when launching  the internet browser and returning to the program.
We overrode EventResume and EventSuspend functions of LEventDispatcher in CMac5LApp to do this and then call the inherited version.

We also deleted an unnecessary call to HideMenuBar found in CPlayerView.

We added ShowMenuBar to the DoBrowse command in CCard.cp. We cannot rely on EventSuspend to do this correctly - see comment in code.

Changes by Elizabeth Hamon, okayed by Eric.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@27098 b24aca02-f418-0410-9c49-835e4aa915c3