Created a new file CStreamTests.cpp which has tests for CStream. These tests are run by running TestAll project in Mac).  These tests should  pass (see note in file for exceptions) before any changes to CStream are made.
Current version of CStream passes all tests, although tests regarding whitespace are designed to meet current setup. The way whitespace is dealt with is a mess currently, as the tests illustrate. When CStream is fixed for this,  the test cases regarding this in CStreamTests need to be updated. 

Changes by Elizabeth and Eric, okayed by Eric.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@27135 b24aca02-f418-0410-9c49-835e4aa915c3