Basic style-run information.
*sigh* I've thrown out a lot of code today, implementing and refactoring
various parts of styled text drawing.  I need to overhaul parts of the
current design to fully integrate styles.

My guess: the 'wchar_t *begin' and 'end' pointers in LineSegment need to be
substantially rethought, and the whole class needs to be properly
encapsulated.  There also needs to be some sort of way to bundle up
character strings and style data, and a way to easily create those bundles
for testing purposes.

Once these issues are addressed, I think the rest of the code should just
fall together.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@27200 b24aca02-f418-0410-9c49-835e4aa915c3