Replacement MD5 library from Aladdin Software.
The old MD5 library was distributed under the GPL, and we're not allowed
to use it in our software.  We *can* use this library.

The old files are no longer available through CVS, even when checking
out the tags of older versions.

The old files are stored in the CVS repository in a file named
REMOVED_FOR_LEGAL_REASONS.tar.gz.  Do not extract the files from that
library without consulting a laywer.