Replace .SEND with SEND
The existing '(obj .send meth-name arg-list)' API is fairly unpleasant to
use.  It's not very Scheme-like, mostly because the entire list of
arguments must always be passed as the final argument to '.send'.  This is
annoying when you have a fixed-length argument list.  Even worse, it's also
less flexible than 'apply', because you must jam all your arguments into
the final 'arg-list'.

The new API is an ordinary Scheme function named 'send', which can be used
with or without 'apply':

  (send obj meth-name arg1 arg2...)
  (apply send obj meth-name arg1 arg2 arg-list)

This change was accomplished by renaming the internal 'send%' to 'send',
exporting it, and converting all the code that called '.send' over to the
new system.

Bug-ID: 2620


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@28986 b24aca02-f418-0410-9c49-835e4aa915c3