Work around %text% "Trying to break line in the middle of a character" bug
When updating (.text) from "" to a non-empty string, we frequently get an
error: "Trying to break line in the middle of a character".  This is
because we've updated our .text *before* our .shape, and there's not enough
space to draw even a single letter.

The correct fix: Split .invalidate and .draw into two separate phases, and
update the drawing architecture to call .draw only at well-defined times,
not during property updates.

Unfortunately, doing this would be non-trivial, so we work around it by a
quick-and-dirty hack.  Specifically, we check to see whether the width of
our .shape has been fully updated yet, and only draw if it has been.

Bug-ID: 2977


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29018 b24aca02-f418-0410-9c49-835e4aa915c3