Fix Scintalla off-by-one error when lexing |#
In Scintilla's lexing loop, it keeps track of i, ch and chNext.  On
each loop iteration, it increments i, sets ch to chNext, and sets
chNext to the character at position i+1.  When we read two characters
in a row, we increment i an extra time to account for that fact, but
we were forgetting to also update chNext so ch would be set properly
on the next iteration.

Bug-ID: 3600


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29028 b24aca02-f418-0410-9c49-835e4aa915c3