Remove last remaining reference to 5L from test/Runtime/halyard
This patch removes all remaining references to 5L in
test/Runtime/halyard, except in code that hasn't run in years
(extract-docs.pl, tool.ss).  It takes a few different approaches.  In
some cases, references have been renamed to Halyard.  In others,
they've just been changed to engine or just the bare word for the
concept referred to.  And 5l-log is renamed to app-log.

Bug-ID: 1856


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29041 b24aca02-f418-0410-9c49-835e4aa915c3