Replace outdated LICENSE.txt and README in halyard/test
The LICENSE.txt inside of halyard/test was useless.  The build process
copies halyard/LICENSE.txt into the binary and test releases, so it
should be sufficient.  This patch also adds a test/README.txt that
should help new users be able to run the program, and adds a note to
LICENSE.txt about where to find source if you have downloaded a -bin
or -test distribution.

Bug-ID: 3699


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29042 b24aca02-f418-0410-9c49-835e4aa915c3