Use SYNTAX-PARAMETERIZE to 'bind' SELF and SUPER instead of MAKE-SELF
Replaces use of MAKE-CAPTURE-VAR/ELLIPSIS to bind SELF and SUPER with
SYNTAX-PARAMETERIZE, which allows there to just be a single top-level
binding which is parameterized at syntax expansion time with the
appropriate SELF and SUPER.

This allows us to replace many of the SYNTAX-CASE expanders that were
simply used to capture self with MAKE-SELF with much simpler and
easier to understand SYNTAX-RULES expanders.  It also means that the
only places that need special knowledge of SELF are METHOD (to
parameterize the syntax appropriately), APP~ (to pick up appropriate
syntax location information when using an implicit SELF), and
DEFINE-NODE (which needs to check at expansion time whether SELF
has a valid expansion).

We still use MAKE-SELF in APP~ to pick up source location information
when SELF is implicit (see also the change to MAKE-CAPTURE-VAR/ELLIPSIS
to preserve source location information).

We also need to patch Swindle to not export its incompatible
definition of DEFINE-SYNTAX-PARAMETER.

For more information on syntax parameters, see the blog post
explaining the basic idea, and the documentation for stxparam.ss:
http://blog.plt-scheme.org/2008/02/dirty-looking-hygiene.html
http://download.plt-scheme.org/doc/360/html/mzlib/mzlib-Z-H-43.html

Bug-ID: 3870


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29053 b24aca02-f418-0410-9c49-835e4aa915c3