Proxy initializer keywords and methods to child elements
We want to create a "captioned movie" class, which contains a movie element
and a caption element.  We want to automatically pass most initializer
keywords and methods directly through to the underlying movie.

This patch implements a simple API for this kind of proxying:

  (define-class %captioned-movie% (%box%)
    (initialize-and-proxy-for-child-element 'foo %foo%)
    (setup
      (.add-child-initializer! :message "Hello, world!")
      (.create-child-element)))

We also add a few other goodies:

  symbol->keyword        Convert 'foo to :foo
  setter-name?           Return true for symbols of form 'set-foo!
  setter-name->keyword   Convert symbols of form 'set-foo! to :foo
  rect-shape             Eqv to (move-rect-left-top-to r (point 0 0))

There is an example implementation in media.ss, which suffers somewhat
because of the rect/bounds-shape-at inconsistencies.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29054 b24aca02-f418-0410-9c49-835e4aa915c3