Add tools/checkfiles, which performs style checks
We're going to want to run this script from somewhere in our pre-commit
hook.  Right now, it checks for consistent line endings: Either CRLF or LF
is good (at least until we switch the tree to one or the other), but the
line endings need to be consistent on a per-file basis.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29055 b24aca02-f418-0410-9c49-835e4aa915c3