Merge r28729: Updater: Initial Vista work
NOT YET TESTED:
* Added a manifest to UpdateInstaller, with requireAdministrator flag
set.
* Added a CommandLine::ExecAsync function which calls ShellExececute
internally, so that we can use UAC privilege escalation.
* Removed all the DIR-WRITABLE? code. It appears to be unwise to believe
the return values on Vista, so I'm not going to leave it around to
tempt people.
* Moved the Updates directory into the user's script data directory,
where the user is guaranteed to have write privileges.
* We no longer attempt to copy UpdateInstaller.exe directly into program
files, because we're not guaranteed to have sufficient privileges.
Instead, we make a copy in a staging area, and let it copy itself in
once it has obtained the necessary privileges.

(cherry picked from commit 36a543ee7dc296be492fcc363a2bea77ecd43760)
Author: emk


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/branches/0.2-stable@29099 b24aca02-f418-0410-9c49-835e4aa915c3