Allow interacting with unit test elements manually
Adds %element-test-case% and %element-test-suite% classes for testing
elements.

%element-test-case% creates a temporary parent element in SETUP-TEST,
uses that as the default element parent in the test method iteself,
and deletes it during TEARDOWN-TEST.  This is a replacement for the
TEST-ELEMENTS form that does the same all at once during the test body,
which doesn't allow execution of those phases to be broken up.

%element-test-suite% adds an interface for interacting with elements
created as part of an %element-test-case%.  By default, it simply runs
all of the tests as normal; however, it has an extra interface that
allows you to run just the SETUP-TEST and test method of each test
individually, allowing you to interact with any elmeents created
to test interactive bahavior manually.

Bug-ID: 3028


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29151 b24aca02-f418-0410-9c49-835e4aa915c3