Allow whitespace-only strings to be drawn into 0-width boxes
The line-breaking algorithm must handle two cases.  When the text is broken
into natural segments (at whitespace, hyphenation points, etc.), either:

 1) the largest segment fits on a single line, or
 2) the largest segment needs to be chopped up further.

We handle case 1 correctly, but we will sometimes unnecessarily raise an
error in case 2.

Case 1: Imagine that we try to draw the string "foo bar" into a
3-character-wide box (assuming fixed-width strings).  This string breaks
into three natural segments: "foo", " ", "bar".  We can put "foo" on the
first line, discard the space, and put "bar" on the second line.  We know
that it's OK to throw away the space, because seg.discardAtEndOfLine is
true.

Case 2a: Imagine that we try to draw the string "abcd" into a
2-character-wide box.  Here, we can break "abcd" into "ab" and "cd", and
lay it out across two lines.

Case 2b: Imagine that we try to draw the string "a" into a 0-character-wide
box.  Here, there's no way to fit the text into the box, so we raise an
error.  For better or for worse, this is currently the desired behavior,
because it tends to catch bugs.

Case 2c: Imagine that we try to draw the string " " into a 0-character-wide
box.  Here, we *should* succeed, because we're at the end of the line, and
seg.discardAtEndOfLine is true, so we can just throw away the space.  But
the current code treats this like case 2b, which is wrong.

This patch adds a check for case 2c, and a variety of unit tests to prove
that the new code works.

Bug-ID: 4250


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29178 b24aca02-f418-0410-9c49-835e4aa915c3