Merge r29066: Lex selection before auto-indenting
When selecing text that is not visible on the screen and indenting,
Scintilla has not yet lexed that text for syntax coloring and
indentation information.  This means that if you press CTRL-A/TAB in
order to re-indent the whole file, it doesn't have enough information
to properly indent it, so the indenting becomes screwed up.  This patch
calls Colourise() to force the lexer to scan the whole selection when
indenting a selection.

(cherry picked from commit 24302f50406758e955527da444f260354f1a51ba)
Author: brian
Bug-ID: 3211


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/branches/0.2-stable@29186 b24aca02-f418-0410-9c49-835e4aa915c3