First working version: Fix broken code backported from trunk
When backporting changes to *.ss code, I accidentally tested the wrong
Runtime directory.  This meant that a number of bugs slipped through: Calls
to exn:fail?, path->string, assert-equal, and other functions that only
exist in newer engines.

This patch finishes the cleanup work, and gives us an engine which
successfully loads VTRA and runs the updater test suites.


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/branches/0.2-stable@29214 b24aca02-f418-0410-9c49-835e4aa915c3