mizzen: Move ruby-objects-test.ss into mizzen
Move all of the engine-indepent tests of the object system into mizzen. 
The tests for Swindle still working had to be rewritten because they 
depended on types from types.ss, which isn't moving to mizzen.  One of 
the tests relied on elements to reproduce a bug, so it was left behind.

Bug-ID: 3939


git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29249 b24aca02-f418-0410-9c49-835e4aa915c3