editor: Start lexing at line boundaries when indenting
When doing auto-indentation, we need to ensure that the region being
indented is lexed, so the indenter will have the information it needs.
We were simply lexing the current selection before indenting, but it
turns out that the lexer is only guaranteed to work when starting
after a newline, so instead we expand the range to lex to the beginning
and end of the start and end lines of the selection.

Bug ID: 4374

git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29309 b24aca02-f418-0410-9c49-835e4aa915c3