mizzen: Update fix-prologues.pl to understand Scheme files
fix-prologues.pl is used to add copyright notices to the top of all
source files.  This patch updates it to detect Scheme files, use
appropriate comment delimiters in Scheme files, and apply the LGPL to
them rather than the GPL.  It also detects whether it's operating on
files in the mizzen directory, and uses "Mizzen" as the name of the
program when in that directory.

Furthermore, when we're in the Mizzen directory, we use 2006-2008 as
the copyright years, the reason being that 2006 is the first time that
any of the files in Mizzen were ever released.  We treat Halyard and
Mizzen as distinct projects, so while the copyright years 1993-2008
encompass everything distributed in Halyard, anything that is part
of Mizzen has existed only as of 2006.

Bug ID: 3939

git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29313 b24aca02-f418-0410-9c49-835e4aa915c3