mizeen: Rename language.ss to mizzen.ss
It makes more sense for this to be called mizzen.ss, now that we have a
name for the language.

Also, removes the inaccurate description of what mizzen.ss does, left
over from when it used to be called lispish.ss, and replaces it with a
very brief description.

Bug ID: 3939

git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29316 b24aca02-f418-0410-9c49-835e4aa915c3