mizzen: Use relative paths instead of library paths in includes
It seems to generally be preferred to use relative paths instead of
library paths in PLT Scheme collections, since it means that it can be
distributed as a PLaneT package, or as a regular collection that goes
into your collection path.  For now, we're just distributing it as a
regular collection, but we may create a PLaneT package later.

Bug ID: 3939

git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@29317 b24aca02-f418-0410-9c49-835e4aa915c3