Move libraries into git submodules
These libraries used to be part of the main git repository, but they caused
a number of problems:

  1) Most of these libraries won't be needed on non-Windows platforms.
  2) The monolithic source tree made Cygwin git almost as slow as Subversion.
  3) There was no easy way to maintain vendor branches for each library.

Therefore, the libraries were edited out of the history of the main Halyard
repository, and given repositories of their own.  This new structure will
make day-to-day development much faster, and allow us to upgrade to the
latest version of PLT cleanly.

To get a local copy of the libraries, run:

  git submodule init
  git submodule update

If you need to modify a file that has been moved to a submodule, the
correct procedure is:

  1) Test your changes.
  2) Commit the submodule.
  3) Push the submodule.
  4) Commit Halyard.
  5) Push Halyard.

In particular, the 2 -> 4 ordering and the 3 -> 5 ordering are critical.
If you forget the first, you'll fail to commit submodule changes.  If you
forget the second, other developers won't have access to the submodule.

We'll also need to update the build system and a fair bit of our workflow.
Please let me know if you have any questions!