Benchmarking library for mizzen
This library provides a simple declarative syntax for benchmarks, some
standard benchmarks for object-model performance, and a command-line driver
program.

  (define-class %simple-dispatch% ()
    (def (do-nothing)
      #f))

  (define-benchmark "Simple method dispatch" 1000000
    (define obj (%simple-dispatch% .new))
    (benchmark (obj .do-nothing)))

Here, the '(define obj ...)' form is one-time setup code, and the
'benchmark' form contains the actual code which we will run 1,000,000
times.

All declared benchmarks can be run and displayed as follows:

  (foreach [report (run-benchmarks)]
    (printf "~a ~a~n" (report .microseconds) ((report .benchmark) .name)))

Current benchmarking results, with all times in microseconds for a single
iteration, and (useless) memory usage deltas that should improve with PLT
v372:

  121000.0 -32768 (collect-garbage)
  0.012 0 Do nothing
  0.106 0 PLT: hash-table-get
  0.117 0 PLT: hash-table-get default
  0.462 0 PLT: hash-table-get default with closure
  0.12 0 PLT: hash-table-put!
  1.826 0 Simple method dispatch
  3.148 0 Calling super
  2.49 0 Read attr
  2.27 0 Write attr
  35.7 0 Create class without attributes
  754.0 4096 Create class with many attributes
  2.7 0 instance-of? (1 level)
  14.7 0 instance-of? (6 levels)

This library will be used as the basis for future mizzen optimizations.

v2:
  Many more benchmarks
  Run tests incrementally
  Attempt to measure memory usage (unsuccessfully in PLT v360)
  Modified to use mizzen.ss instead of language.ss