Make CommonHeaders.h compile with GCC
(Not tested with MSVC++.)

A number of MSVC++-specific idioms have crept into our codebase over the
years.  This patch fixes a number of problems, including some non-portable
explicit template instantiations and some missing empty virtual
destructors.

We also add a stub to Makefile.am that can be used to explicitly generate a
precompiled header.  For now, this is just for testing--it doesn't generate
a .deps entry, and it isn't integrated into the build process.