Common: Some unit tests now pass on MacOS X
[Not tested with MSVC++.  Changes to TestAll.cpp not committed.]

Some of the Common library now compiles on MacOS X, and--if you comment out
portions of TestAll.cpp--the unit tests will actually pass.

  Old-Style ImlUnit Tests
  Ran 225 tests, 0 failed, 100% passed
  OK

  New-Style TestCase Tests
  OK: 23 passed, 0 skipped

The actual hacked-up version of TestAll.cpp isn't included in this commit,
because it would represent a non-trivial regression to the Windows codebase
and it would keep us from merging all the other changes so far back to the
master branch.

The biggest obstacles at this point appears to be (1) anything Scheme
related, and (2) TValue.  The latter relies heavily on implicit
conversions, some of which turn out to be ambiguous in standard C++.  It's
been too many years since we built with anything but MSVC++.

General:
  Many tiny g++-compatibility tweaks
  Replaced last uses of DISABLE_COPY_AND_ASSIGN with boost::noncopyable
  Constructors for TPoint, TRect and TPercent redesigned to prevent
    unwanted implicit conversions

TException:
  Convert from deprecated ostrstream to ostringstream

FileSystem:
  SetBaseDirectory now converts the base directory to an absolute path

TValue:
  Inlined TValue::FindType to prevent duplicate copies when we link

TestCase:
  std::exception can't be used with a string argument