Build fix: Replace preprocessor error in CommandLine.cpp with exception
Somehow, my last commit broke the non-Windows build.  As far as I can tell,
I had some stale automake state and never tried to build CommandLine.cpp.

We used to issue a compiler error for platforms without ExecAsync, but we
don't need it unless we're trying to make the updater work.  And that's a
very long-term goal anywhere but Windows.