Common: Make TestAll run on MacOS X with some tests disabled
This is the version of TestAll.cpp which I've had in my tree for the last
several commits.  I've modified it to disable unported test cases using
using preprocessor macros instead of comments, so we that don't disrupt the
MSVC++ build.

Also, we haven't had a TString class in many years, so I deleted a
commented-out test case referring to it.