Rakefile: remove tag AS=x.y.z command
In Git, we just tag by doing git tag vx.y.z followed by git push --tags.
This is simple enough that we don't need a rake command to do it.  If
someone forgets to push the tag, the build script should fail early in
the process and so be easy to fix.