Replace Ogg Theora test card with "Duck and Cover" version
The new test card uses the Ogg Theora version of "Duck and Cover", and has
a button which will take you to the XiphQT site, where you can download the
codecs.  Unfortunately, XiphQT is giving me pretty serious problems under
Windows Vista.  We're definitely not ready to use this in a production
engine yet.