Replace implicit TValue conversions with tvalue_cast<T>(...)
We use TValue to store dynamically-typed data.  But to use that data, we
must first convert it to an ordinary C++ type.  We used to perform these
conversions using implicit conversion operators:

  TValue v(...);
  std::string s(v);

Unfortunately, this led to a rat's nest of possible implicit conversion
paths between any two types, and tended to make our code highly dependent
on bugs in MSVC++'s ambiguity resolution algorithm.  Needless to say, this
made it more-or-less impossible to compile TValue with any other C++
compiler.

The new API uses explicit cast operators, by analogy with dynamic_cast<T>
and the other standard C++ casting operators:

  TValue v(...);
  std::string s(v);

I ran all the unit tests and walked through CommonTest carefully by hand,
and the engine appears to be solid.  But please keep your eyes open for new
bugs, because this is a fairly major patch.

With a bit of luck, this code will be more likely to work with g++ than
what we had before.