g++: Specialize tvalue_cast<T> from within Halyard namespace
[Not tested with MSVC++.]

We can now build TValue.cpp with g++.  This was pretty straightforward.
All the operator overloading ambiguities were eliminated by
d504b25fc7efee47e74a2345167db03802077e3c, and we just needed to make sure
that we performed the template specializations in the right namespace.

Now that this major obstacle is out of the way, porting the rest of Common
should be much easier.