Fix nasty GC leak in TSchemePtr.h
This may be one of the dumbest and most expensive bugs in the history of
the engine, and it dates back to my very first TSchemeInterpreter.h commit.

We have a carefully-designed class which wraps Scheme pointers and makes
sure that the GC doesn't collect them until we're ready.  Unfortuntely, the
wrapper class doesn't have a destructor, and therefore it never actually
releases pointers back to the GC.

Unfortunately, fixing this bug does not--by itself--fix our memory leak
problem.  According to Eli, we may also want to try using custodians to
wrap our sandboxed namespaces, or even try upgrading to v372 and using his
sandbox module.