Common: Build lang/scheme and re-enable all test cases
I _think_ everything in Common is now building under MacOS X (I need to
alphabetize the file names to be sure).  If you want to use the official
release of PLT Scheme, make sure it's in a directory with no space
characters anywhere in the path, and tell ./configure where to find it:

  ./configure --with-mzscheme=/Applications/Development/PLT-v360

You will also need to set up a symlink allowing the mzscheme framework to
be found at runtime:

  cd /Library/Frameworks
  ln -s /Applications/Development/PLT-v360/lib/PLT_MzScheme.framework .

I don't know whether this will work with MacPorts mzscheme.  If it doesn't,
feel free to send patches.