Untested: Rewrite TSchemePtr to use scheme_malloc_immobile_box
[Until further notice, assume that the build is hopelessly broken on all
platforms.  Even if it _compiles_, it certainly won't survive the first
GC.]

The original TSchemePtr class used scheme_dont_gc_ptr to prevent pointers
from being collected.  But with the new 3m collector, this no longer works,
because pointers may be moved.  Instead, we allocate an immovable box on
the Scheme heap, and store the pointer in there.

This API is reminiscent of the old Mac OS 'Handle' type, which was also a
fixed pointer to a box containing a movable pointer.

Fortunately, we can preserve the old API, which is always nice.  Smart
pointers are generally a win.