buildfile.rb: Don't include compiled directories in SVN checkin or ZIP
When we moved from Subversion to Git, we ported buildscript to use svn
add to tag the Runtime and Fonts directories instead of svn cp.  But,
since we're adding fresh directories, we don't have the svn:ignore
properties set up, so we were adding the compiled directories as well.

This patch moves the build process into a different directory,
bin_dir, and then copies the Runtime and Fonts directories out of the
pristine source tree when tagging, so we don't have to deal with any
build products.