Add buildscript as a submodule
We've been relying on people having a separate checkout of buildscript
in their Ruby include path, which means that people doing builds have
had to manually keep sync between the version of buildscript and the
version of Halyard.  This patch adds buildscript as a submodule in
the tools directory, and uses that version in buildfile.rb.