Add an experimental C++ API for interfacing with PLT's precise GC
We add two new classes to Common/lang/Scheme.

  TSchemeReg: An exception-safe version of MZ_GC_REG and friends
  TSchemeArgs: Convenient local Scheme_Object* argc/argv support

These two classes work together to implement a (somewhat) more convenient
API for dealing with the precise GC from C++.  They include a fair number
of assertions, and we can add more as needed to reduce our manual proof
burdern.  This replaces the API described in "Inside PLT MzScheme":

http://download.plt-scheme.org/doc/372/html/insidemz/insidemz-Z-H-3.html#node_sec_3.1

Frankly, the PLT precise GC API is miserable to use, either with or without
our wrappers.  I examined the PLT source to see what they did, and was
surprised to discover that they don't use the precise GC API at all.
Instead, they rely heavily on 'mzc --xform' to automatically add the
necessary source code annotations.  They even do this for the C++ code in
MrEd, which is surprising, because 'mzc --xform' has weak C++ support.

Unfortunately, this isn't a viable strategy for us, because we need
exception-handling support (and it would be really annoying to call 'mzc
--xform' from our MSVC++ build process).  I suppose that we could use 'mzc
--xform' normally, and replace the macros it generates with exception-safe
versions.  Perhaps we'll reconsider this strategy in the future if events
warrant.

For examples of using the new API, see the comments in the headers.  Also
included is a written code review policy for any code using this API.