Explain GC safety assumption made by TSchemePtr's constructor
As it turns out, some call sites using the implicit TSchemePtr constructor
are only safe if we assume that scheme_malloc_immobile_box cannot cause a
GC.  This is true in PLT v372.  Here, we document that assumption, and
explain what to do if future versions of PLT change the implementation of
scheme_malloc_immobile_box.