Begin converting code to use new GC API
This is a very rough port: It compiles, but it has never been tested or
run.  It's extremely likely that it still contains bad GC-related bugs
caused by mistakes in converting to the new API.  I've done some code
review, but not the full checklist specified in GC-REVIEW-CHECKLIST.txt.

I have an ulterior motive for committing this code now: I want a record of
any mistakes that I discover from this point on, so that we have written
institutional knowledge of just how hard this API is to get right.

One major piece of work remains: The FindBucket function assumes that it
can compare TValue pointers and use them as keys in a std::map.  Since the
new GC moves pointers in memory, this is no longer safe (or even
minimally functional).