FindBucket: Use stable identifiers with std::map, not movable pointers
FindBucket now uses the Env and Module enumerations to represent all
possible environments and modules, because we can't rely on stable pointers
to the underlying Scheme objects across a garbage collection.

Of course, this means that we _must_ discard all BucketKey objects when
shutting down a TInterpreter, because it invalidates anything found in
SCRIPT_ENV.  Fortunately, the relevant code already exists:

  TSchemeInterpreter::~TSchemeInterpreter() {
      // We don't actually shut down the Scheme interpreter.  But we'll
      // reinitialize it later if we need to.
      sBucketMap.clear();
  }

We're still not able to run TestAll, so none of this code has been tested
yet.  But we're getting close.