Remove in-tree copies of standard PLT collections
In the past, we've had two copies of the standard PLT collections:

  libs/plt/collects
  test/Runtime (also contains mizzen and halyard)

We've actually been using the copies in test/Runtime.  But this presents
several problems:

  1) When building with an out-of-tree version of PLT (for example, on the
     Mac), we want to use the collections actually shipped with PLT,
     because they're guaranteed to match the mzscheme version we built
     against.
  2) Re-importing all the collections every time we upgrade PLT bulks up
     our main Halyard repository for no good reason.
  3) If we use the collections in lib/plt/collects, this makes it much
     easier to maintain local patches that deal with Win32-specific issues,
     because we have the full power of git merges available to us when
     upgrading PLT.

Note that we have a list of local patches to the PLT collections in
Common/lang/scheme/plt-v360-patches.txt.

I've discussed this change with Brian, and he thinks it's a good idea.
We'll need to make some further changes to loader.ss, to the release
process, and to how we run the in-tree copies of Halyard and TestAll on
Windows.