Look for PLT collections in standard locations
As part of our plan to stop relying on copies of standard PLT collections
found in test/Runtime/{mzlib,...}, we now search for PLT collections in all
the places specified by (find-library-collection-paths).  This is how the
standalone mzscheme does it.

With this change applied, it's finally possible to run TestAll and load
loader.ss.  But when trying to load anything from Scripts, it fails with
the following assertion:

  ERROR: ASSERTION FAILURE: TPrimitives.cpp:323: inName != ""

It looks like the debugging has officially begun.