Document difference between SCHEME_SYM_VAL and SCHEME_BYTE_STR_VAL
The former returns an interior pointer, and requires special handling.  The
latter returns an ordinary pointer.  This is documented in a small note at
the bottom of Section 3.1 of "Inside PLT MzScheme", which is a bit obscure.
Brian suggested that I mention it explicitly in the GC review checklist,
which is an excellent idea.