Fix symbol-name lookup bug ('make check' now runs!)
The previous version of this code attempted to use SCHEME_BYTE_STR_VAL to
access the content of a Unicode string, with predictably unsuccessful
results.  We now convert it to a byte string before attempting to call
SCHEME_BYTE_STR_VAL.

With this patch, 'make check' now runs at 100% under MacOS X.  We still
need to do fine-grained code review of the changes in commit
513161048bcdf1e9337a33d2e6323e17eb56181d, and we need to spend some time
running the unit tests from a stress-testing harness.  Nonetheless, this is
an important milestone.