Search for standard PLT library collections in several places
Now that the PLT library collections have been moved out of Runtime (at
least in the build tree), we need to search in several places to actually
find them:

  1) Runtime.  If we're using a released engine, the collections will still
     appear here.
  2) The system "collects" directory, if any.  The collections will appear
     here if we're using the system's copy of mzscheme.
  3) ../libs/plt/collects.  The collections will appear here if we're being
     run from the halyard/test directory in the build tree.

Note that we do all the path-name manipulation in Scheme (because our
FileSystem library is too weak, and boost::filesystem breaks precompiled
headers under MSVC++), but we actually call set_collects_path from C++
(because I can't find a Scheme version of this API).

This means that we need to allow the SchemeSetCollectsPath primitive to be
called before we set up sScriptEnv.