PLT v372: Rudimentary support for building with MSVC++
To build this version of Halyard with MSVC++, see the instructions in
HACKING.txt about running 'rake libs'.  All the test cases pass, and I
exercised 3/4ths of halyard/test without seeing any problems.

Note that the diffs for the *.sln and *.vcproj files are fairly large,
because Visual Studio decided to change the line endings and shuffle some
other information around.