HACKING.txt: Correct build instructions
Asking to build the solution instead of run it is the correct way to
build PLT scheme.

Note: You must delete all subdirectories of test/Runtime except for
halyard and mizzen in order to run the program.  There is a bug
somewhere in mzscheme (or in our new code) that causes the engine to
crash if there are leftover empty collection directories in there.