Restore support for trusting precompiled *.zo files
We originally added this code in 9867ff7b7dc3d05686732d9eaca18b45d4c25265,
and disabled it in 2514df68d0a2db609781914842185b76a58d04f9 for the upgrade
to PLT v372.

Now that we've patched v372 to support always-treat-zo-and-so-as-newer, we
can re-enable this code.  But this time, we actually test whether
always-treat-zo-and-so-as-newer is actually available, because we want to
support building Halyard with standard, unmodified versions of PLT.