paths: Update node full names to include leading slashes
We are beginning to rebuild the path system.  In the new path system,
we would like to be able to refer to absolute paths beginning with a /.
It would be confusing if paths could begin with a / in the path, but
the node name itself doesn't contain a path.

This patch changes every node's full name to include a leading /, and
updates everything to be able to deal with that.  It does not add
support to the path syntax to recognize paths with leading slashes.

Porting-Notes:
* All card, sequence, and group declarations must begin with a leading
  slash.  The following regular expression (emacs syntax) may help:
  (\(card\|sequence\|group\)\(\s-+\)\([-a-z0-9][-a-z0-9/]*\)
  -> (\1\2/\3
* Anywhere that refers to any card, sequence, or group by the variable
  name instead of by a path must be updated to include the leading /.
  There is no regular expression that can catch all of these, but a
  variation of the above may help for a limited subset, and others can
  be found by just loadingg the program and fixing all of the
  undefined identifier errors.