Bug fix: Clear bucket map properly after a failed reload
This bug was triggered by e9ddc4302c1e389412687bc4daffd29368554189, and
first noticed in e14bb5ff0fad9c856a4c5bbe38942c75bbd36499. But the roots of
this bug go much deeper in the history of Halyard--all the way back to
4207615eb2e16a832a701dfd3ca0fa9353f1c875 in 2003.

The symptoms: When loading the script, we would frequently see an error
of the form "Expected a color, got #<TValue: (uninitialized)>" the first
time we called DefStyle.  Brian noticed the pattern: This error would occur
if and only if the _previous_ attempt to load the script had failed.

I added a check in a97279014c3bbd94f68c17617f098565002a15be to catch the
uninitialized TValue objects as soon as they were created.  (The
decision to return uninitialized TValue objects here dates to
9efbf0ac6443f6074ff3fe9f1ee2603f91affcca, which I reviewed and approved.)

Once I added the check, the problem became clear: We weren't clearing
sBucketMap properly after a failed attempt to load a script.  This meant
that functions such as "color?" and "point?" would persist in sBucketMap
between interpreter instances.  But since these functions are only good for
the lifetime of the interpreter, the stale versions in sBucketMap would
always return #f when called, causing SchemeStructToTValue to fail, and
to return an uninitialized TValue.

Before e9ddc4302c1e389412687bc4daffd29368554189, this didn't matter,
because sBucketMap was keyed with the actual Scheme_Env and module
pointers.  So even though we kept a stale copy of "color?" in sBucketMap,
there was no way to recover it.  (Thank you to Brian for this explanation
of why this bug went hidden so long.)  But when we replaced the pointer
keys with an enumeration, we revealed the underlying bug.

There were an annoying number of different mistakes leading up to this bug,
all of them mine:

 1) sBucketMap was cleaned up in TSchemeInterpreter::~TSchemeInterpreter.
    But when the constructor threw an exception, we never had a
    fully-constructed TSchemeInterpreter, and we therefore never called the
    destructor.  The fix: Make sBucketMap a member variable, so that C++
    automatically destroys it when TSchemeInterpreter's constructor fails.
 2) TInterpreterManager::Run was lacksidasical about cleaning up
    TInterpreter objects, so we weren't necessarily calling TInterpreter's
    destructor when we should.

The underlying problem here is that TSchemeInterpreter contained a
ridiculous number of static members that should have been non-static.  This
was part of an ill-conceived attempt to make TSchemeInterpreter::CallScheme
convenient to call from other classes.  Again, the fault was mine.

This patch cleans up a significant amount of cruft in TInterpreter.cpp and
TSchemeInterpreter.cpp.

The detailed changes included in this patch are:

  * Modified TInterpreterManager::Run to not leak TInterpreter objects
    when reloading a script fails.
    - We no longer store a pointer to the interpreter in TInterpreterManager;
      it has to use the singleton lookup interface like everyone else.
  * Modified TSchemeInterpreter to clean up bucket map properly.
    - Made sBucketMap a member variable, so that it gets cleaned
      up properly if we throw an exception out of our
      constructor.  The C++ standard guarantees that member
      variables will properly destroyed if a constructor fails,
      but offers no equivalent guarantee for static member
      variables.
  * Made most other static members of TInterpreter non-static.
    - Moved point, rect, etc., creation functions into TSchemeConv.
    - Moved idle handling into TInterpreterManager
      - Renamed SchemeIdle -> Idle, and moved to TCommonPrimitives.cpp
      - TSchemeInterpreter::DoIdle -> TInterpreterManager::DoIdle